rebasing fixes

+++ pkg/controller/infrastructure/infraflow/reconcile.go
@@ -476,6 +476,9 @@ func (fctx *FlowContext) ensureSSHKeyPair(ctx context.Context) error {
 }
 
 func (fctx *FlowContext) ensureShareNetwork(ctx context.Context) error {
+	if fctx.config.Networks.SubnetID != nil {
+		return fctx.ensureShareNetworkForExistingSubnet(ctx)
+	}
 	if fctx.config.Networks.ShareNetwork == nil || !fctx.config.Networks.ShareNetwork.Enabled {
 		return fctx.deleteShareNetwork(ctx)
 	}
@@ -519,5 +522,36 @@ func (fctx *FlowContext) ensureShareNetwork(ctx context.Context) error {
 	}
 	fctx.state.Set(IdentifierShareNetwork, created.ID)
 	fctx.state.Set(NameShareNetwork, created.Name)
+	return nil
+}
+
+// ensureShareNetworkForExistingSubnet ensures the shared network for an existing subnet. Because the subnet may be shared among many different shoots,
+// it could be that there is already a sharednetwork associated with a subnet. This function is responsible for detecting the shared network associated with the subnet.
+func (fctx *FlowContext) ensureShareNetworkForExistingSubnet(_ context.Context) error {
+	networkID := ptr.Deref(fctx.state.Get(IdentifierNetwork), "")
+	subnetID := ptr.Deref(fctx.state.Get(IdentifierSubnet), "")
+	current, err := findExisting(fctx.state.Get(IdentifierShareNetwork),
+		"",
+		fctx.sharedFilesystem.GetShareNetwork,
+		func(_ string) ([]*sharenetworks.ShareNetwork, error) {
+			list, err := fctx.sharedFilesystem.ListShareNetworks(sharenetworks.ListOpts{
+				NeutronNetID:    networkID,
+				NeutronSubnetID: subnetID,
+			})
+			if err != nil {
+				return nil, err
+			}
+			return sliceToPtr(list), nil
+		})
+
+	if err != nil {
+		return err
+	}
+
+	if current != nil {
+		fctx.state.Set(IdentifierShareNetwork, current.ID)
+		fctx.state.Set(NameShareNetwork, current.Name)
+	}
+
 	return nil
 }
